+++ src/Common/MPI/debug.hpp
@@ -105,7 +105,7 @@ boost::this_thread::sleep(boost::posix_time::milliseconds(msec));               
 **/
 #define PEDebugVector(v,length) { \
   std::cout << ::CF::Common::mpi::PE::instance().rank() << "/" << ::CF::Common::mpi::PE::instance().size() << ": " << #v << " " << length << " ( " << std::flush; \
-  for(int _tmp_i_=0; _tmp_i_<(const int)(length); _tmp_i_++)  std::cout << v[_tmp_i_] << " "; \
+  for(int _tmp_i_=0; _tmp_i_<length; _tmp_i_++)  std::cout << v[_tmp_i_] << " "; \
   std::cout << " )\n" << std::flush; \
 }
